requirements.txt
Flask==2.0.1
Pillow==10.1.0
# Flask-Mail==0.9.1

